[reorg] Begin moving GPOS into the new directory layout.
https://github.com/harfbuzz/harfbuzz/commit/c7307ca06
