+++ src/hb/ot/layout/GPOS/pos_lookup.rs
@@ -1,0 +1,33 @@
+use ttf_parser::GlyphId;
+use crate::hb::ot_layout::LayoutLookup;
+use crate::hb::ot_layout_common::PositioningLookup;
+use crate::hb::ot_layout_gsubgpos::Apply;
+use crate::hb::ot_layout_gsubgpos::OT::hb_ot_apply_context_t;
+
+impl LayoutLookup for PositioningLookup<'_> {
+    fn props(&self) -> u32 {
+        self.props
+    }
+
+    fn is_reverse(&self) -> bool {
+        false
+    }
+
+    fn covers(&self, glyph: GlyphId) -> bool {
+        self.coverage.contains(glyph)
+    }
+}
+
+impl Apply for PositioningLookup<'_> {
+    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
+        if self.covers(ctx.buffer.cur(0).as_glyph()) {
+            for subtable in &self.subtables {
+                if subtable.apply(ctx).is_some() {
+                    return Some(());
+                }
+            }
+        }
+
+        None
+    }
+}
