+++ src/hb/ot/layout/mod.rs
@@ -1,1 +1,2 @@
 mod GSUB;
+mod GPOS;
